+++ client/src/main/kotlin/su/plo/voice/client/audio/source/JitterBuffer.kt
@@ -2,46 +2,20 @@ package su.plo.voice.client.audio.source
 
 import su.plo.voice.proto.packets.tcp.clientbound.SourceAudioEndPacket
 import su.plo.voice.proto.packets.udp.clientbound.SourceAudioPacket
-import java.util.*
-import java.util.concurrent.LinkedBlockingQueue
-import java.util.concurrent.PriorityBlockingQueue
 
-class JitterBuffer(
-    private val packetDelay: Int
-) {
+interface JitterBuffer {
 
-    private val queue: Queue<PacketWithSequenceNumber> = if (packetDelay <= 1) {
-        LinkedBlockingQueue()
-    } else {
-        PriorityBlockingQueue(
-            packetDelay * 2,
-            compareBy { it.sequenceNumber }
-        )
-    }
+    fun offer(packet: SourceAudioPacket)
 
-    private var endPacket: SourceAudioEndPacket? = null
+    fun offer(packet: SourceAudioEndPacket)
 
-    fun offer(packet: SourceAudioPacket) {
-        if (endPacket != null && packet.sequenceNumber > endPacket!!.sequenceNumber) {
-            this.endPacket = null
-        }
+    fun poll(): PacketWithSequenceNumber?
 
-        queue.offer(SourceAudioPacketWrapper(packet))
-    }
+    fun isEmpty(): Boolean
 
-    fun offer(packet: SourceAudioEndPacket) {
-        this.endPacket = packet
+    fun isNotEmpty(): Boolean = !isEmpty()
 
-        queue.offer(SourceAudioEndPacketWrapper(packet))
-    }
-
-    fun poll(): PacketWithSequenceNumber? {
-        if (endPacket != null || queue.size >= packetDelay) {
-            return queue.poll()
-        }
-
-        return null
-    }
+    fun reset() {}
 
     sealed interface PacketWithSequenceNumber {
         val sequenceNumber: Long